split out date and date interval filters
